+++ src/redturtle/prenotazioni/browser/static/widget/js/fields/SelectField.js
@@ -5,28 +5,44 @@ import WidgetContext from '../utils/widgetContext';
 
 // import './index.less';
 
-const SelectField = ({ value, id, row, vocId, multi, customUpdateField }) => {
-  const { vocabularies, updateField } = useContext(WidgetContext);
+const SelectField = ({
+  value,
+  id,
+  row,
+  vocId,
+  multi,
+  customUpdateField,
+  key,
+  placeholder,
+}) => {
+  const { vocabularies, updateField, getTranslationFor } = useContext(
+    WidgetContext,
+  );
   const vocab = vocabularies[vocId];
   if (!vocab) {
     return '';
   }
   const options = vocab.items.map(item => {
-    return { value: item.token, label: item.title };
+    return { value: item.token, label: getTranslationFor(item.title) };
   });
-  const selectValue = options.filter(option => {
+  let selectValue = options.filter(option => {
     if (Array.isArray(value)) {
       return value.includes(option.value);
     } else {
       return value === option.value;
     }
   });
+  if (!multi) {
+    selectValue = selectValue[0] || null;
+  }
   return (
     <Select
       isMulti={multi ? true : false}
       isClearable={true}
-      defaultValue={selectValue}
+      value={selectValue}
       options={options}
+      key={key}
+      placeholder={placeholder}
       onChange={option => {
         let newValue = null;
         if (Array.isArray(value)) {
@@ -48,6 +64,8 @@ SelectField.propTypes = {
   multi: PropTypes.bool,
   row: PropTypes.number,
   customUpdateField: PropTypes.func,
+  key: PropTypes.any,
+  placeholder: PropTypes.string,
 };
 
 export default SelectField;